+++ taqueria-vscode-extension/package.json
@@ -27,14 +27,27 @@
     "ecadlabs"
   ],
   "activationEvents": [
-    "onCommand:taqueria.helloWorld"
+    "onCommand:taqueria.init"
   ],
   "main": "./out/extension.js",
   "contributes": {
+    "configuration": {
+      "title": "Taqueria",
+      "properties": {
+        "taqueria.path": {
+          "type": [
+            "string",
+            "null"
+          ],
+          "default": null,
+          "description": "Specifies taqueria binary path"
+        }
+      }
+    },
     "commands": [
       {
-        "command": "taqueria.helloWorld",
-        "title": "Hello World"
+        "command": "taqueria.init",
+        "title": "Taqueria: Init"
       }
     ]
   },
